backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
APScheduler==3.10.4
requests==2.31.0
cryptography==41.0.7
python-telegram-bot==20.7
beautifulsoup4==4.12.2
python-dotenv==1.0.0

